feat: update air data properties to use double types and parse values from JSON

+++ lib/domain/entities/air_data.dart
@@ -1,7 +1,7 @@
 class AirData {
   final double temperature;
   final double humidity;
-  final int gasPpm;
+  final double gasPpm;
 
   AirData({
     required this.temperature,
